requirements.txt
uwsgi==2.0.21
coreapi==2.3.3
django==2.2.28
django-autocomplete-light==3.8.2
django-cors-headers==2.4.0
django-filter==2.4.0
djangorestframework==3.11.2
djangorestframework-simplejwt==4.4.0
paramiko==2.7.2
psycopg2-binary==2.8.6
python-cinderclient==7.4.1
python-novaclient==17.7.0
python-glanceclient==3.6.0
scp==0.14.4
django-admin-auto==0.1.1
django-simpleui==4.0.3
ambari==0.1.7
